+++ FitnessApi/Controllers/FoodController.cs
@@ -16,17 +16,26 @@ namespace FitnessApi.Controllers;
 public sealed class FoodController
 {
     private readonly IFoodService _foodService;
+    private readonly IAuthService _authService;
 
-    public FoodController(IFoodService foodService)
+    public FoodController(IFoodService foodService, IAuthService authService)
     {
         _foodService = foodService;
+        _authService = authService;
     }
 
     [FunctionName("GetMacros")]
     public async Task<IActionResult> GetMacros(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userId = Guid.Parse(req.Query["userId"]);
 
         var user = await _foodService.GetUserMacros(userId);
@@ -36,9 +45,16 @@ public sealed class FoodController
     
     [FunctionName("AddCustomMacros")]
     public async Task<IActionResult> AddCustomMacros(
-        [HttpTrigger(AuthorizationLevel.User, "post", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+        
         string requestBody;
         using (var streamReader = new StreamReader(req.Body))
         {
@@ -57,9 +73,16 @@ public sealed class FoodController
 
     [FunctionName("AutocompleteFood")]
     public async Task<IActionResult> AutocompleteFood(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var query = req.Query["query"];
 
         var user = await _foodService.AutocompleteFood(query);
@@ -69,9 +92,16 @@ public sealed class FoodController
 
     [FunctionName("GetRecentUserFoods")]
     public async Task<IActionResult> GetRecentUserFoods(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userId = Guid.Parse(req.Query["userId"]);
         var date = DateTime.Parse(req.Query["date"]);
 
@@ -82,9 +112,16 @@ public sealed class FoodController
 
     [FunctionName("SearchFood")]
     public async Task<IActionResult> SearchFood(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var query = req.Query["query"];
         var page = int.Parse(req.Query["page"]);
 
@@ -95,9 +132,16 @@ public sealed class FoodController
     
     [FunctionName("GetFoodV2")]
     public async Task<IActionResult> GetFoodV2(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var foodId = long.Parse(req.Query["foodId"]);
 
         var foods = await _foodService.GetFoodById(foodId);
@@ -107,9 +151,16 @@ public sealed class FoodController
     
     [FunctionName("GetUserFoodV2")]
     public async Task<IActionResult> GetUserFoodV2(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userFoodId = long.Parse(req.Query["userFoodId"]);
 
         var foods = await _foodService.GetUserFoodV2(userFoodId);
@@ -119,9 +170,16 @@ public sealed class FoodController
     
     [FunctionName("GetAllUserFoodV2")]
     public async Task<IActionResult> GetAllUserFoodV2(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userId = Guid.Parse(req.Query["userId"]);
         var date = DateTime.Parse(req.Query["date"]);
 
@@ -132,9 +190,16 @@ public sealed class FoodController
     
     [FunctionName("AddUserFoodV2")]
     public async Task<IActionResult> AddUserFoodV2(
-        [HttpTrigger(AuthorizationLevel.User, "post", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         string requestBody;
         using (var streamReader = new StreamReader(req.Body))
         {
@@ -152,9 +217,16 @@ public sealed class FoodController
     
     [FunctionName("UpdateUserFoodV2")]
     public async Task<IActionResult> UpdateUserFoodV2(
-        [HttpTrigger(AuthorizationLevel.User, "put", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         string requestBody;
         using (var streamReader = new StreamReader(req.Body))
         {
@@ -170,9 +242,16 @@ public sealed class FoodController
     
     [FunctionName("QuickAddUserFoodV2")]
     public async Task<IActionResult> QuickAddUserFoodV2(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userId = Guid.Parse(req.Query["userId"]);
         var foodId = long.Parse(req.Query["foodId"]);
         var date = DateTime.Parse(req.Query["date"]);
@@ -184,9 +263,16 @@ public sealed class FoodController
     
     [FunctionName("QuickRemoveUserFoodV2")]
     public async Task<IActionResult> QuickRemoveUserFoodV2(
-        [HttpTrigger(AuthorizationLevel.User, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userId = Guid.Parse(req.Query["userId"]);
         var foodId = long.Parse(req.Query["foodId"]);
         var date = DateTime.Parse(req.Query["date"]);
@@ -198,9 +284,16 @@ public sealed class FoodController
     
     [FunctionName("DeleteUserFoodV2")]
     public async Task<IActionResult> DeleteUserFoodV2(
-        [HttpTrigger(AuthorizationLevel.Function, "delete", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userFoodId = long.Parse(req.Query["userFoodId"]);
         
         await _foodService.DeleteUserFoodV2(userFoodId);
@@ -210,9 +303,16 @@ public sealed class FoodController
     
     [FunctionName("GetCurrentUserMacrosV2")]
     public async Task<IActionResult> GetCurrentUserMacrosV2(
-        [HttpTrigger(AuthorizationLevel.Function, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var userId = Guid.Parse(req.Query["userId"]);
         var date = DateTime.Parse(req.Query["date"]);
 
@@ -223,9 +323,16 @@ public sealed class FoodController
     
     [FunctionName("SearchFoodByBarcode")]
     public async Task<IActionResult> SearchFoodByBarcode(
-        [HttpTrigger(AuthorizationLevel.Function, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+
         var barcode = req.Query["barcode"];
 
         var user = await _foodService.GetFoodByBarcode(barcode);
@@ -234,7 +341,7 @@ public sealed class FoodController
     }
 
     [FunctionName("MaliciousComplianceHTTP")]
-    public async Task MaliciousComplianceHttp([HttpTrigger(AuthorizationLevel.Function, "get", Route = null)]
+    public async Task MaliciousComplianceHttp([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
     {
         await _foodService.RefreshCashedFoodDb(log);
