+++ FitnessApi/Controllers/UserController.cs
@@ -15,12 +15,43 @@ namespace FitnessApi.Controllers;
 public sealed class UserController
 {
     private readonly IUserService _userService;
+    private readonly IAuthService _authService;
 
-    public UserController(IUserService userService)
+    public UserController(IUserService userService, IAuthService authService)
     {
         _userService = userService;
+        _authService = authService;
     }
 
+    [FunctionName("Auth")]
+    public async Task<IActionResult> Auth(
+        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "auth")]
+        HttpRequest req,
+        ILogger log)
+    {
+        string requestBody;
+        using (var streamReader = new StreamReader(req.Body))
+        {
+            requestBody = await streamReader.ReadToEndAsync();
+        }
+
+        var data = JsonConvert.DeserializeObject<Auth>(requestBody);
+
+        if (data == null)
+        {
+            return new BadRequestResult();
+        }
+
+        var auth = await _userService.AuthByEmailPassword(data.Email, data.Password);
+
+        if (auth.Item1 == false)
+        {
+            return new UnauthorizedResult();
+        }
+        
+        return new OkObjectResult(auth.Item2);
+    }
+    
     [FunctionName("AuthV2")]
     public async Task<IActionResult> AuthV2(
         [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)]
@@ -120,9 +151,16 @@ public sealed class UserController
 
     [FunctionName("GetUser")]
     public async Task<IActionResult> GetUser(
-        [HttpTrigger(AuthorizationLevel.Function, "get", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = null)]
         HttpRequest req, ILogger log)
-    { 
+    {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+        
         var userId = Guid.Parse(req.Query["userId"]);
 
         await _userService.UpdateLastLogin(userId);
@@ -157,9 +195,16 @@ public sealed class UserController
     
     [FunctionName("UpdateUserSex")]
     public async Task<IActionResult> UpdateUserSex(
-        [HttpTrigger(AuthorizationLevel.User, "post", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)]
         HttpRequest req, ILogger log)
-    { 
+    {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+        
         string requestBody;
         using (var streamReader = new StreamReader(req.Body))
         {
@@ -182,9 +227,16 @@ public sealed class UserController
     
     [FunctionName("UpdateUserPaid")]
     public async Task<IActionResult> UpdateUserPaid(
-        [HttpTrigger(AuthorizationLevel.User, "put", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+        
         string requestBody;
         using (var streamReader = new StreamReader(req.Body))
         {
@@ -207,9 +259,16 @@ public sealed class UserController
     
     [FunctionName("UpdateUserUnits")]
     public async Task<IActionResult> UpdateUserUnits(
-        [HttpTrigger(AuthorizationLevel.User, "post", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+        
         string requestBody;
         using (var streamReader = new StreamReader(req.Body))
         {
@@ -232,9 +291,16 @@ public sealed class UserController
     
     [FunctionName("DeleteUser")]
     public async Task<IActionResult> DeleteUser(
-        [HttpTrigger(AuthorizationLevel.User, "delete", Route = null)]
+        [HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = null)]
         HttpRequest req, ILogger log)
     {
+        var authed = await _authService.CheckAuth(req);
+        
+        if (authed == false)
+        {
+            return new UnauthorizedResult();
+        }
+        
         var userId = Guid.Parse(req.Query["userId"]);
 
         await _userService.DeleteUser(userId);
